Add support for multiple images per integration
Until now it was only possible to have an `icon.png` in a package. But having documentation with dashboards example in the future, multiple images per integrations are needed. This adds support for `.png` and `.jpg` images per package.

The path to the images changes because of this and is now `/img/{integration-name}-{version}/{img}`, before it was `/img/{integration-name}-{version}.png

As an example:

* Before: `/img/envoyproxy-0.0.5-.png`
* After: `/img/envoyproxy-0.0.5/icon.png`