Add support for .tar.gz files and change download path
Previously only .zip files were exposed. Now also .tar.gz files are exposed. It should be decided if long term we only need one format but this will allow us to play around with both formats for now.

This commit also changes the download API. Previously it was: `/package/{package}/get`. This is now changed to `/package/{package}.tar.gz` and `/package/{package}.zip`. It feels more logical to me to have the suffix already in the URL.

To make such changes less breaking for the UI in the future, the download link is also added to the list output:

```
[
  {
    "description": "This is the envoyproxy integration.",
    "download": "/package/envoyproxy-0.0.2.tar.gz",
    "icon": "/img/envoyproxy-0.0.2.png",
    "name": "envoyproxy",
    "version": "0.0.2"
  },
  {
    "description": "This is the envoyproxy integration with improved features.",
    "download": "/package/envoyproxy-0.0.5.tar.gz",
    "icon": "/img/envoyproxy-0.0.5.png",
    "name": "envoyproxy",
    "version": "0.0.5"
  }
]
```

It seems some of the information that is above in the list package output should also be in the detail package output. I will tackle this later.