Addition of easy generation of example integration packages (#25)
Instead of manually adding more packages, this adds a script that allows to generate packages based on a config file. With this, its much easier to simulate how it will work with many integrations, integrations with several versions. It also helps to keep the structure of each integration up-to-date as it all comes from one template and can be regenerated at any time.

The `.zip` and `.tar.gz` files were removed from the repository and also the `packages` directory. When now running it locally these must be generated first manually through `mage build`. This creates a package directory under `build/packages`. If built with the docker container, these are generated directly.

Three additional integrations (apache, mysql, rabbitmq) were added in this PR as example but it's planned to follow up with a list of about 12 integrations. The content inside each integration is just a duplicate of the template so these cannot be used for actual usage.

The previous envoyproxy package was moved under `testdata/packages` and is only used for testing now. It also serves as example package.