Add iptables integration
This adds the iptables integration to the registry. It's a copy of the iptables Filebeat modules with all the existing assets.

For now the raw configs for Filebeat are included. There is no way to use it directly at the moment but were added to make it possible for us to experiment with it.

The fields.yml was added and a template must be generated out of it. It only contains the fields from iptables but not the ECS and Filebeat fields. The `index.asciidoc` file was copied over with the original content. More work is also needed here.

The magefile was updated to support adding actual implementations from the example directory. Later on, these will be in a different repository but for simplicity are in here now.