Introduce static file handling
As discussed in https://github.com/elastic/integrations-registry/issues/53, some of the files can be generated statically and don't require it's own handler. Pregenerating some of the static content can also simplify the code. To be able to work with both methods in parallel and evaluate this new approach, there is a catchAll handler introduce which is called when no other handler matches. Then it is checked if a file for the request URL exists and if yes, it is served. Like this the deployment model of the services stays exactly the same for now.

To make the above possible, the generated packages are now moved to `public/package` directory. This is done by `mage build`. The side affect of having all packages in there is that also all files inside the package can be directly statically served.

As a first step, the static handler for the `.tar.gz` downloads and `/info` endpoint was removed.