Proposal: List screenshots in manifest (#49)
Discussing with Henry some of the designs, it turned out that we will need in the future support for multiple screenshots, potentially even quick intro videos. As these screenshots should also have a description, I propose the following addition to the `manifest.yml` file:

```
screenshots:
- name: overview-logs.png
  title: This shows the overview of the logs dashboard.
- name: overview-metrics.png
  title: Metrics Dashboard.
- name: getting-started.mp4
  title: Getting started with the envoyproxy integration.
```

All the screenshots must be listed in here and will be shown in the order listed. Videos can also be added and the description is in here. The path to each screenshot / video will be based on a convention.

I don't suggest we add a list of screenshots or videos in our first iteration but just have one, but already set the technical foundation for it. This means for now our implementation would look as following:

```
screenshots:
- name: overview-logs.png
  title: This shows the overview of the logs dashboard.
```

This PR does not change yet the existing integrations and implementation of the registry, only the example manifest.