Rename /list endpoint to /search (#57)
In recent discussions it became more obvious to me that our current `/list` endpoint is more a search endpoint for packages. It returns a list of packages based on the search query params. At the moment these queries are versions of the different services but long term it could also be a text search. By default the endpoint returns the list of all packages with the most recent version.

This is a breaking change!