Change path of icons
This changes the current icon path from `/img/{package}-{version}/icon.png` to `/package/{package}-{version}/img/icon.png`. The reasoning is that directly the assets from the package itself are served so no special handlers are needed in the backend. The other breaking change is that in case no icon exists, a 404 is returned instead of the default icon. It's up the integrations manager to decide what to do for a default icon.

With https://github.com/elastic/integrations-registry/issues/45 the API response for icons will be refactored. For now I left the API response the same, meaning it contains the icon path even if no icon exists, but will change that when tackling #45.