Removing package generating code
The code to generate packages was removed and instead the generated packages are commited. For the next step of testing the UI and the integrations registry we need very different types of packages to test the edge cases. This is not possible with generated packages as they are too different from each other and also not the actual packages. Because of this, the current generated packages were taken and commited into the repo. Now these can be used for small experiments to see how variations work and still have a list of integrations. From the integrations-manager perspective, nothing changes.