Add /categories endpoint (#81)
The `/categories` endpoint is to serve information about how many packages per category exist. An example output looks as following:

```
[
  {
    "id": "logs",
    "title": "Logs",
    "count": 16
  },
  {
    "id": "metrics",
    "title": "Metrics",
    "count": 14
  }
]
```

For now the output does not support any filtering. It just takes all the newest packages and adds them to the categories. The category titles are hardcoded in the registry code for now. Later on we should validate the titles of each package.

The decision was made for the more complex output already know, so we can add additional fields in the future without breaking the API.

In the future I imagine this endpoint supports the same filtering capabilities as the search endpoint.

Closes https://github.com/elastic/integrations-registry/issues/75