Adjust description that fields.yml should be used for generation (#89)
To keep all the flexibility with the fields.yml also after creating a package, for now I suggest that the integrations managers converts it to index templates. This has the advantage that we do not need to merge all fields.yml into one file in the package and can delay the decision to the installation if one or multiple templates (logs vs metrics) are created.