Introduce package filtering in /search endpoint (#88)
To show which versions of a package exist, the `/search` endpoint must support filtering by a package and then return all versions of this package.

This introduces the parameter `package` with which queries like `/search?package=mysql` can be created.

These results can be used for showing a drop down when upgrading a version to show which other versions exist.

To have an example package for testing this, the `multiversion` package was added. It can be queried with `/search?package=multiversion`.