Generate root index endpoint instead of having it hardcoded
The `index.json` file in the root directory was the only file shipped as part of the repo instead of being generated. This is changed with this PR. It lays the foundation to dynamically read in the version in the future. Also formatting wise it's guaranteed that all json docs have the same output.

The field `service.name: "integrations-registry"` was added in this PR. This is to have an example on how more data can be added but is not used at the moment.