Prepare release 0.1.0 (#102)
This is the first release that should be tagged. So far the changelog was not used but that will change now. All development will still happen on top of master.

All files were updated for the this release. Breaking changes might still happen until we hit 1.0 but they will be documented.