Remove not needed files in Docker image
Currently in the docker images all packages existed twice, in the `dev` directory which is need to build the packages and then under `public`. At the moment this does not amake a big difference but in the future with many packages, this would duplicate the size of the container. Because of this as a last step, the dev directory is removed.

I also took this opportunity to remove a few more files from the Docker image which are only used for testing or CI. Like this we ship an image with only the relevant files. In the future we might remove even more files like the Golang code and only ship the binary. But I keep this for a later point.