Fix Docker image to specific Golang version (#107)
So far the `:latest` image was used for Golang. To be used in production I prefer to tied it to a very specific version so we know what is run and what we test is identical.

A file `.go-version` was introduced which long term should serve as the single point of through for the go version, also for CI. At the moment it is only there for documentation purpose.

The ENTRYPOINT of the Dockerfile was split up into ENTRYPOINT to make it easier to overwrite parameters.

Note: As part of this change I also looked into using the alpine Golang image but it does not have Git included, which makes working with the Golang mod dependency trickier.