Adding concept of input to package
In the future metricbeat / filebeat and the agent will only support inputs. With this inputs become a first class citizen in our stack. An input is basically an agent configuration + an ingest pipeline. At the moment the package content is focused on having a config for the Beat or agent and all pipelines in one place. This complicates 2 things:

* Knowing which ingest pipeline belongs to a specific input
* Building integrations with multiple inputs: https://github.com/elastic/integrations/pulls

Having the concept could simplify things as the package builder must not try to prevent naming configs of ingest pipeline by introducing extra long names. Also it should simplify testing as often testing is focused on inputs. With this all assets related to an input are together.

As part of this PR there is an example on how such an input structure could look like. This should not replace the old place of ingest pipelines. If a user wants to build a package with just a ingest pipeline but not an input, this should also be possible in the future.

The changed structure is described in the ASSET.md file.