Remove packages.path config and replace it with public_dir (#118)
The only thing needed to start the server is having it pointed to the public directory to serve files from there. What the internal name is for the packages directory is an implementation detail.

Some refactoring of the code was needed for this change. Now path variables are passed to the handler instead of using a global variable. This also makes it simpler to test individually instead of modifying global variables.