Remove assets form generated packages
The registry so far contained generated packages. These packages contained all the same assets and were mainly designed to test the overview page of the package manager. But this now caused problems as many dashboards for example had the same ids. Because of this, all the assets are removed from these example packages and only the package with the icons are now left which is enough to show a nice home screen.