Add base package
The base package contains all the assets, which are used by all packages. This package is expected to be installed before any other package is installed. EPM should install it automatically and no user interaction should be needed.

For now this only contains the two default ILM policies for logs and metrics. More assets could be added later.

A flag `internal: true` was added to the packages so these packages should not be shown in the overview. By default the internal packages are not exposed in the `/search` endpoint. To also get internal packages, the query param `?internal=true` has to be set. In the `/categories` endpoint, internal packages are not counted.

Further changes:

* Test package for testing `internal: true` added
* README updated with new endpoint.
* Add test for `internal=true` query param