Fix test data generation (#130)
When updating the generation process of assets, the test packages were not kept up to date. Some refactoring was needed to make it available again. Now it is always run as part of `mage check` so it is run on all tests. Like this, the test packages are also always up-to-date.