Prepare renaming of integration-registry to package-registry
We decided some time ago to rename the integration-registry to package-registry as it contains more then just integrations. This is a first step to rename the service by adjusting the docs. A second PR will be needed when renaming the repository to adjust import golang paths.

Part of https://github.com/elastic/integrations-registry/issues/113