Adjust API output for Kibana compatibility version
Previously the json output for the compatible Kibana version looked as following:

```
"kibana": {
  "version.min": "7.0.0"
}
```

As this causes troubles on the JS side, it is changed to:

```
"kibana": {
  "version": {
    "min": "7.0.0"
  }
}
```

This should make it easier to parse.

To make it for package creators still possible to use the short notation in the manifest file, go-ucfg is used to read the manifest file so we get this functioanlity for free.

An additional change which happened in this PR is that if a max or min version are empty, it is not part of the API output anymore.

Closes https://github.com/elastic/integrations-registry/issues/44