Add format_version to manifest file
The field `format_version` is added to each manifest to indicate which package version format is used. This will become useful if in case one day we will make breaking changes to the structure and can then on the EPM side make a smart decision on how to process the package.

Additional changes:

* All the config reading had to be updated to `config` from `yaml` as it uses ucfg. This should have done before but was missing. The problem only showed up now because of different names of variable and field name.