Add version ranges instead of min / max version (#139)
Currently it is only possible to specify a min and max version as compatibility. This has its limitations as the especially the max version is not always known in advance but all versions of the minor should be compatible.

Implementing this change I realised we should only support ranges or min / max but mixing them gets confusing. Having just one string to define the full range also becomes powerful. For example 6.8 was released after 7.0 and it could be that a package is compatible with 6.8 and 7.2. This is not possible to represent at the moment.

My proposal in this PR is to remove min / max and instead use a `version` or `versions` string.

```
requirement:
  kibana:
    versions: ">=7.1.0 <8.0.0"
```

The supported formats by the library I use can be found here: https://github.com/blang/semver#ranges