Cleanup ASSETS.md file, improve docs
Some parts in the ASSETS.md file have gotten out of date recently. This PR intends to clean up the ASSETS.md file. One big part here is to introduce a `definition` package which contains all the fields direclty with additional documentation. This makes it possible that we also always tests this package and make sure it is up-to-date with the code.