Update test data generation with 2 directories
During packaging assets are modified. Because of it must be tested if the input and output is as expected. So far the generated data was written into the same directory as the test packages. Now this does not work anymore because the files are modified. This PR moves the generated data for the testdata into a public directory. This will also help with having nice diffs on what is the new content that is generated.