Fix fields.yml which still had a key (#172)
Some of the fields.yml still had a `- key: foo` prefix as it was copied over from the Beats where this is used for the docs. But this breaks the template generation in Kibana. In the future, we should also validate fields.yml files to make sure they are compatible.

In addition, fields.yml for the system package were added.