Add download and path urls (#174)
At the moment, only the `/search` endpoint contains the download url. This is now also added to the package info. In addition a field `path` is added which is the path to the package info. This should help with breaking changes in the future so EPM can rely on this information to find the info about a package instead of having to hard code it.