Prefix package download url with /epr/{package-name} (#169)
Prefixing it with `/epr/{package-name}` simplifies the download stats rules. Only the `.tar.gz` packages are move to a different url. An example of a download path is: `/epr/apache/apache-1.0.1.tar.gz`.

The new download path is available in the package output. It is a breaking change but I would expect that from a package manager point of view it should just work as we rely on the download field.