Cleanup nginx package
This cleans up the nginx package by adding the missing fields like title, ingest pipeline etc.

In addition, the field title is made a requirement to make sure it is always set.

The ecs.yml field is added to the logs part to have the ecs fields around for now. Long term the fields that are actually used have to be added.