Validate dataset to check for pipeline config (#188)
This PR checks a dataset manifest to make sure the config is as expected. If no pipeline is specified, it checks if pipeline files are there. If yes, it checks if a `default.json` or `default.yml` are there and takes it as the ingest pipeline. If files are there with different names but no default, it is an invalid dataset and an error is throw. The goal is to make sure, only valid packages are shipped.

Additional changes:

* Move dataset code to its own file
* Use `Validate` magic from go-ucfg
* Introduce package to test for default pipeline behaviour
* Add check that if no Kibana version is specified, it returns true by default