Remove docs from assets which are out-of-date (#199)
Some of the docs entry in the ASSETS.md file are by now inaccurate and not needed anymore. Definitions will be part of the Contributing guide and assets for the existing beats not supported. More cleanup will be needed for each assets but this removes the overall parts which are not needed.