Dashboard exporting tool
To build the current packages, I used some tooling code on my end to select the correct saved objects from Kibana and decode them. In this PR I want to share the code I use.

The code is a bit messy as it is copied out of the project I initially built in Golang to load assets but now all happens in EPM. This simplifies the code here but still needs clean.

The code needs in input yaml file with all the assets defined (apache.yml as an example) and then a target dir to be specified which is normally the package. The command then looks like the following:

```
go run . -target=/path/to/my/package -input="./package-registry/dev/tools/kibana-exporter/exports/apache.yml"
```

Publishing as a draft as needs cleanup.