Add basic logs and metrics template
For logs and metrics indices, we will provide default templates. This makes sure, that all indices which are created in logs-*-* and metrics-*-* follow the convention.

In this PR, a first example of these templates is provided. In total there are 4 templates, one for mapping and one for settings for each. The content of the templates itself is very basic at the moment. It contains a few fields from ECS and defaulting to keyword for all fields otherwise if it is a string. More work must be put into these templates in the future.

These templates are a bit different from the other fields.yml as they are installed globally and don't depend on a data source. Because of this they are in the top level Elasticsearch directory and directly in the json format. From an installation perspective in the package manager my expectations are as following:

* EPM takes the name of the file and the content and puts it directly into Elasticsearch (no processing)