Add streams as part of Datasource / Inputs
The streams are defined in the dataset and reference an input. The input is defined in the datasource. For the UI to not have to combine to two together, the API already provides the datasource with all inputs and streams listed inside.

This change also extended the test packages to have better examples for this change