RFC: Have datasource type to link to correct solution
Some of the datasources will not be configured by the ingest manager UI. An example for this are Endpoint or Uptime. The UI still shows the datasource entry and must know which solution to link to when someone selects such a datasource. For this, a field `solution: endpoint` was introduced inside a datasource.