Remove generated packages and enabled beats packages (#286)
In the beginning, there were a few generated packages to fill the screen in EPM. With the Beats module migration script kicked off, we now have more then enough packages avaiable. Because of this the existing generated packages are removed and the Beats packages are enabled during the check and build process.

There are some example packages which overlap with the Beats packages like system or nginx and it is at the moment important that we keep using the manual built packages for testing because they contain all the details. This is not a problem as all the Beats packages have version 0.0.1. As long as the manual built packages have a newer version, things will just work.