Change default for /search endpoint with package filter
So far when a `/search?package=foo` query was run, by default all versions of the package found were returned. This was in contrast to that by default only the most recent version of a package is returned. To make the API less surprising the behavior was changed that by default only the most recent package is returned.

As there might be still a need to return all packages, and additional param `all=true` was added. This can now be used in combination with all query params. To get the same result as before, `/search?package=foo&all=true` has to be run. By default all is false.

add test for all

add one more all test