Fix System Host overview dashboard
On the Beats side, we took this dashboard and replace the hostname field with the local hostname. This had the nice onboarding experience that a user had the data directly filtered on his own host. This does not work anymore so this filter has to be removed. Instead I expect us when linking for example from an agent to a dashboard or similar, that we can apply some filters in advance.