Add independent Dockerfile (#367)
The package storage and the package registry are soon going to be decoupled. To be able to still build a docker container with both inside a Dockerfile is needed that can pull in the dependencies from an external sources instead of local files.

To get this started, a separate Dockerfile is introduced that does not load any files from the repository directly but pulls it from Github. This is to start play around with this idea.

The end goal is being able to pass in the version of the packages that should be used to built the container and the version of the registry.

Where this Dockerfile should land in the end is not clear yet but the package-registry seems to be a good place at the moment.